Do not push system resources from cloud to edge
